wtp/webrtc/RTCPeerConnection-setRemoteDescription-replaceTrack.https.html

<!doctype html>
<meta charset="utf-8">
<title>RTCPeerConnection.prototype.setRemoteDescription - replaceTrack</title>
<script src="/resources/webrtc-ie-shim.js"></script><script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="RTCPeerConnection-helper.js"></script>
<script>
  'use strict';
  
  // The following helper functions are called from RTCPeerConnection-helper.js:
  //   getUserMediaTracksAndStreams
  
  var _slicedToArray = function () { function sliceIterator(arr, i) { var _arr = []; var _n = true; var _d = false; var _e = undefined; try { for (var _i = arr[Symbol.iterator](), _s; !(_n = (_s = _i.next()).done); _n = true) { _arr.push(_s.value); if (i && _arr.length === i) break; } } catch (err) { _d = true; _e = err; } finally { try { if (!_n && _i["return"]) _i["return"](); } finally { if (_d) throw _e; } } return _arr; } return function (arr, i) { if (Array.isArray(arr)) { return arr; } else if (Symbol.iterator in Object(arr)) { return sliceIterator(arr, i); } else { throw new TypeError("Invalid attempt to destructure non-iterable instance"); } }; }();
  
  async_test(function (t) {
    var caller = new RTCPeerConnection();
    return getUserMediaTracksAndStreams(2).then(t.step_func(function (_ref) {
      var _ref2 = _slicedToArray(_ref, 2),
          tracks = _ref2[0],
          streams = _ref2[1];
  
      var sender = caller.addTrack(tracks[0]);
      return sender.replaceTrack(tracks[1]).then(t.step_func(function () {
        assert_equals(sender.track, tracks[1]);
        t.done();
      }));
    })).catch(t.step_func(function (reason) {
      assert_unreached(reason);
    }));
  }, 'replaceTrack() sets the track attribute to a new track.');
  
  async_test(function (t) {
    var caller = new RTCPeerConnection();
    return getUserMediaTracksAndStreams(1).then(t.step_func(function (_ref3) {
      var _ref4 = _slicedToArray(_ref3, 2),
          tracks = _ref4[0],
          streams = _ref4[1];
  
      var sender = caller.addTrack(tracks[0]);
      return sender.replaceTrack(null).then(t.step_func(function () {
        assert_equals(sender.track, null);
        t.done();
      }));
    })).catch(t.step_func(function (reason) {
      assert_unreached(reason);
    }));
  }, 'replaceTrack() sets the track attribute to null.');
  
  async_test(function (t) {
    var caller = new RTCPeerConnection();
    return getUserMediaTracksAndStreams(2).then(t.step_func(function (_ref5) {
      var _ref6 = _slicedToArray(_ref5, 2),
          tracks = _ref6[0],
          streams = _ref6[1];
  
      var sender = caller.addTrack(tracks[0]);
      assert_equals(sender.track, tracks[0]);
      sender.replaceTrack(tracks[1]);
      // replaceTrack() is asynchronous, there should be no synchronously
      // observable effects.
      assert_equals(sender.track, tracks[0]);
      t.done();
    })).catch(t.step_func(function (reason) {
      assert_unreached(reason);
    }));
  }, 'replaceTrack() does not set the track synchronously.');
  
  async_test(function (t) {
    var expectedException = 'InvalidStateError';
    var caller = new RTCPeerConnection();
    return getUserMediaTracksAndStreams(2).then(t.step_func(function (_ref7) {
      var _ref8 = _slicedToArray(_ref7, 2),
          tracks = _ref8[0],
          streams = _ref8[1];
  
      var sender = caller.addTrack(tracks[0]);
      caller.close();
      return sender.replaceTrack(tracks[1]).then(t.step_func(function () {
        assert_unreached('Expected replaceTrack() to be rejected with ' + expectedException + ' but the promise was resolved.');
      }), t.step_func(function (e) {
        assert_equals(e.name, expectedException);
        t.done();
      }));
    })).catch(t.step_func(function (reason) {
      assert_unreached(reason);
    }));
  }, 'replaceTrack() rejects when the peer connection is closed.');
  
  async_test(function (t) {
    var expectedException = 'InvalidModificationError';
    var caller = new RTCPeerConnection();
    return getUserMediaTracksAndStreams(2).then(t.step_func(function (_ref9) {
      var _ref10 = _slicedToArray(_ref9, 2),
          tracks = _ref10[0],
          streams = _ref10[1];
  
      var sender = caller.addTrack(tracks[0]);
      caller.removeTrack(sender);
      // replaceTrack() should fail because the sender should be inactive after
      // removeTrack().
      return sender.replaceTrack(tracks[1]).then(t.step_func(function () {
        assert_unreached('Expected replaceTrack() to be rejected with ' + expectedException + ' but the promise was resolved.');
      }), t.step_func(function (e) {
        assert_equals(e.name, expectedException);
        t.done();
      }));
    })).catch(t.step_func(function (reason) {
      assert_unreached(reason);
    }));
  }, 'replaceTrack() rejects when invoked after removeTrack().');
  
  async_test(function (t) {
    var expectedException = 'InvalidModificationError';
    var caller = new RTCPeerConnection();
    return getUserMediaTracksAndStreams(2).then(t.step_func(function (_ref11) {
      var _ref12 = _slicedToArray(_ref11, 2),
          tracks = _ref12[0],
          streams = _ref12[1];
  
      var sender = caller.addTrack(tracks[0]);
      var p = sender.replaceTrack(tracks[1]);
      caller.removeTrack(sender);
      // replaceTrack() should fail because it executes steps in parallel and
      // queues a task to execute after removeTrack() has occurred. The sender
      // should be inactive. If this can be racy, update or remove the test.
      // https://github.com/w3c/webrtc-pc/issues/1728
      return p.then(t.step_func(function () {
        assert_unreached('Expected replaceTrack() to be rejected with ' + expectedException + ' but the promise was resolved.');
      }), t.step_func(function (e) {
        assert_equals(e.name, expectedException);
        t.done();
      }));
    })).catch(t.step_func(function (reason) {
      assert_unreached(reason);
    }));
  }, 'replaceTrack() rejects after a subsequent removeTrack().');
  
  // TODO(hbos): Verify that replaceTrack() changes what media is received on
  // the remote end of two connected peer connections. For video tracks, this
  // requires Chromium's video tag to update on receiving frames when running
  // content_shell. https://crbug.com/793808

</script>
